Use typed NodeClass in node constructors.

+++ dataset/GitHub_Java/oracle.graal/graal/com.oracle.graal.replacements/src/com/oracle/graal/replacements/nodes/BasicArrayCopyNode.java
@@ -26,9 +26,7 @@ import com.oracle.graal.api.meta.*;
 import com.oracle.graal.debug.*;
 import com.oracle.graal.graph.*;
 import com.oracle.graal.nodeinfo.*;
-import com.oracle.graal.nodes.CallTargetNode.InvokeKind;
 import com.oracle.graal.nodes.*;
-import com.oracle.graal.nodes.java.*;
 import com.oracle.graal.nodes.spi.*;
 import com.oracle.graal.nodes.type.*;
 import com.oracle.graal.nodes.virtual.*;
@@ -36,10 +34,10 @@ import com.oracle.graal.nodes.virtual.*;
 @NodeInfo
 public abstract class BasicArrayCopyNode extends MacroStateSplitNode implements Virtualizable {
 
-    public static final NodeClass<BasicArrayCopyNode> TYPE = NodeClass.create(BasicArrayCopyNode.class);
+    public static final NodeClass<BasicArrayCopyNode> TYPE = NodeClass.get(BasicArrayCopyNode.class);
 
-    public BasicArrayCopyNode(NodeClass<? extends MacroNode> c, InvokeKind invokeKind, ResolvedJavaMethod targetMethod, int bci, JavaType returnType, ValueNode... arguments) {
-        super(c, invokeKind, targetMethod, bci, returnType, arguments);
+    public BasicArrayCopyNode(NodeClass<? extends BasicArrayCopyNode> c, Invoke invoke) {
+        super(c, invoke);
     }
 
     protected ValueNode getSource() {
@@ -91,68 +89,37 @@ public abstract class BasicArrayCopyNode extends MacroStateSplitNode implements 
 
     @Override
     public void virtualize(VirtualizerTool tool) {
-        ValueNode sourcePosition = tool.getReplacedValue(getSourcePosition());
-        ValueNode destinationPosition = tool.getReplacedValue(getDestinationPosition());
-        ValueNode length = tool.getReplacedValue(getLength());
-
-        if (sourcePosition.isConstant() && destinationPosition.isConstant() && length.isConstant()) {
-            int srcPos = sourcePosition.asJavaConstant().asInt();
-            int destPos = destinationPosition.asJavaConstant().asInt();
-            int len = length.asJavaConstant().asInt();
+        if (getSourcePosition().isConstant() && getDestinationPosition().isConstant() && getLength().isConstant()) {
+            int srcPos = getSourcePosition().asJavaConstant().asInt();
+            int destPos = getDestinationPosition().asJavaConstant().asInt();
+            int length = getLength().asJavaConstant().asInt();
+            State srcState = tool.getObjectState(getSource());
             State destState = tool.getObjectState(getDestination());
 
-            if (destState != null && destState.getState() == EscapeState.Virtual) {
+            if (srcState != null && srcState.getState() == EscapeState.Virtual && destState != null && destState.getState() == EscapeState.Virtual) {
+                VirtualObjectNode srcVirtual = srcState.getVirtualObject();
                 VirtualObjectNode destVirtual = destState.getVirtualObject();
-                if (!(destVirtual instanceof VirtualArrayNode)) {
+                if (!(srcVirtual instanceof VirtualArrayNode) || !(destVirtual instanceof VirtualArrayNode)) {
                     return;
                 }
-                if (len < 0 || !checkBounds(destPos, len, destVirtual)) {
+                if (((VirtualArrayNode) srcVirtual).componentType().getKind() != Kind.Object || ((VirtualArrayNode) destVirtual).componentType().getKind() != Kind.Object) {
                     return;
                 }
-                State srcState = tool.getObjectState(getSource());
-                if (srcState != null && srcState.getState() == EscapeState.Virtual) {
-                    VirtualObjectNode srcVirtual = srcState.getVirtualObject();
-                    if (((VirtualArrayNode) destVirtual).componentType().getKind() != Kind.Object) {
-                        return;
-                    }
-                    if (!(srcVirtual instanceof VirtualArrayNode)) {
-                        return;
-                    }
-                    if (((VirtualArrayNode) srcVirtual).componentType().getKind() != Kind.Object) {
-                        return;
-                    }
-                    if (!checkBounds(srcPos, len, srcVirtual)) {
-                        return;
-                    }
-                    if (!checkEntryTypes(srcPos, len, srcState, destVirtual.type().getComponentType(), tool)) {
-                        return;
-                    }
-                    for (int i = 0; i < len; i++) {
-                        tool.setVirtualEntry(destState, destPos + i, srcState.getEntry(srcPos + i), false);
-                    }
-                    tool.delete();
-                    if (Debug.isLogEnabled()) {
-                        Debug.log("virtualized arraycopyf(%s, %d, %s, %d, %d)", getSource(), srcPos, getDestination(), destPos, len);
-                    }
-                } else {
-                    ValueNode source = srcState == null ? tool.getReplacedValue(getSource()) : srcState.getMaterializedValue();
-                    ResolvedJavaType sourceType = StampTool.typeOrNull(source);
-                    if (sourceType == null || !sourceType.isArray()) {
-                        return;
-                    }
-                    ResolvedJavaType sourceComponentType = sourceType.getComponentType();
-                    ResolvedJavaType destComponentType = destVirtual.type().getComponentType();
-                    if (!sourceComponentType.equals(destComponentType)) {
-                        return;
-                    }
-                    for (int i = 0; i < len; i++) {
-                        LoadIndexedNode load = new LoadIndexedNode(source, ConstantNode.forInt(i + srcPos, graph()), destComponentType.getKind());
-                        tool.addNode(load);
-                        tool.setVirtualEntry(destState, destPos + i, load, false);
-                    }
-                    tool.delete();
+                if (length < 0 || !checkBounds(srcPos, length, srcVirtual) || !checkBounds(destPos, length, destVirtual)) {
+                    return;
+                }
+                if (!checkEntryTypes(srcPos, length, srcState, destVirtual.type().getComponentType(), tool)) {
+                    return;
+                }
+                for (int i = 0; i < length; i++) {
+                    tool.setVirtualEntry(destState, destPos + i, srcState.getEntry(srcPos + i), false);
+                }
+                tool.delete();
+                if (Debug.isLogEnabled()) {
+                    Debug.log("virtualized arraycopyf(%s, %d, %s, %d, %d)", getSource(), srcPos, getDestination(), destPos, length);
                 }
             }
         }
     }
+
 }
